+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated admin places tab names to "Places" and "Unverified"
 - Updated AGAM to use average rating if review_count > 0 otherwise fallback to gmaps_rating
 - Updated stopover card aspect ratio to make it taller and skinnier
+- Updated edit duration modal to use a more standard timer picker
 
 ### Fixed
 - Fixed upload button text stays as replace photo even after uploading a new photo